requirements.txt
Flask==1.1.2
Flask==2.3.2
flask_sqlalchemy==3.0.5
PyJWT==1.6.4
PyJWT==2.1.0
PyJWT==2.8.0
SQLAlchemy==1.4.32
